requirements.txt
google-generativeai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
jupyter
nest_asyncio
langchain
langchain-google-genai
beautifulsoup4
lxml
pyautogengoogle-adk[a2a]>=0.1.0
